Remove global: getNextRank
